[audio][drivers] Power down until first client
RingBuffer::SetActiveChannels is optional; audio drivers
must work well for clients who _never_ call it.

That said, for hardware that can be powered-down (and audio
drivers that support RingBuffer::SetActiveChannels for
clients to control the hardware's power), we would ideally
put the hardware into a powered-down state by default, until
the first client connects.

One that occurs, however, we want client changes to the
hardware's power level to be "sticky" and persist even as
RingBuffers are destroyed and others created.

With this CL, during initial configuration (bootup), we
power-down the hardware. Upon the creation of the first
RingBuffer, we automatically power-up. Subsequently, the
only thing that changes the hardware's power level is the
SetActiveChannels() call.

Bug: b/371081956
Bug: b/371088613

Change-Id: I35e53a75c6b3e465314df8586034a0c3331a5627
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130235
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>