[sestarnix] Remove FileSystem root node label workaround for "fuse"
Allow the root node of "fuse" filesystems to be labeled during
resolution of the filesystem's labeling details. Previously this
could result in a deadlock with userspace, if the filesystem was
marked for "fs_use_xattr" labeling, when attempting to fetch the
node's security attribute.  Now that "genfscon" statements in the
policy are respected, the workaround can be removed.

Bug: 369328835
Change-Id: I707bb9239f2079fc9fe77f9d931344dad8f77459
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134252
Reviewed-by: Nelly Vouzoukidou <nellyv@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>