[owners] Remove dworsham@google.com from tools/testgen/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I198a48504618791a557ad2a38fcfd96f39089d95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133331
Commit-Queue: Danny Rosen <dannyrosen@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>