[starnix][aio] Refactor internals of AioContext
This CL moves a bunch of logic from AioContext into simpler objects. For
example, the work of converting between the UAPI opcode and the Rust
enum for the OpType is now part of the OpType implementation.

This change makes it easier to understand the main control flow of the
asynchronous IO operations because all the type conversion is moved out
into leaf types.

Change-Id: I6b3a752cb79245217896a9731dfa13143a4e7f6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133872
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>