[netstack3] Move Mtu to SendOptions
Take the mtu argument from the IpSocketHandler API and move it to the
SendOptions trait. The option is only used in a single place: to limit
the MTU of sent ICMPv6 errors.

Also change the representation in SendIpPacketMeta to use the Mtu type
directly, and default to using a maximum u32 value everywhere. The IP
layer should always be enforcing the device MTU on egress, meaning when
we start doing so we'll apply a min operation between the device and the
requested MTUs.

Bug: 371593049
Change-Id: I95a1a0b9acae7f7e8b17db4f217da73ba68b5ea8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130773
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>