[ffx][config] Removing level options from config set
This removes the `--level` and `--build-dir` options from
`ffx config set` and `ffx config remove`. As a result, this
command only allows writing to the user level configuration.

This is being done to simplify the configuration space by limiting
the writers for each of the levels.

Bug: 371552677

Change-Id: I29078850afa26801679e3f87473e5b21dfa22fbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129875
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>