[kernel][arm64] newer arm cores always have MPIDR.MT set, so tweak how features are printed
A minor tweak that keeps the system from printing some features for
every cpus since modern cpus always have 0 in AFF0 (unless SMT capable).

This particular logic is not really correct, but it keeps the amount of
feature spew down a bit until it can be really refactored to pass the
cluster information into the arch layer such that it truly knows where
the cluster boundaries are.

Also tweaked the types of a few functions nearby to remove some
extraneous casts that were unnecessary.

Change-Id: I1917c203733b7b3a65861ef0a758da6e194ef984
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134815
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>