[kernel][vm] Back split bits w/ a share count
Add a share count field to pages and back split bits with it.

Also add some opaque metadata storage to the compression system. The
caller is able to update the metadata for a page anytime, including
while compression is in-progress or the page is compressed and bound
to a reference. When compression produces a page for the caller (via
Decompress, MoveReference, or Compress failure) it will return the
most updated version of the caller-supplied metadata.

VmCowPages uses this metadata storage to preserve the share count across
Page<->Reference conversions.

No functional change.

Test: compression_unittests
Bug: 42058561
Bug: 338300808
Change-Id: I4c8cda1d2804f34426fa5265ad8fc78a4244ebe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068495
Commit-Queue: David Worsham <dworsham@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>