[coverage][arm64] Test group for arm64 coverage builders
This is a trimmed down group that only includes tests that contribute to
extra source coverage when run on arm64.

We expect using this group to help improve performance and stability of
arm64 coverage builders.

This change only introduces the group, a followup change in infra will
switch arm64 coverage builders to actually use this group.

See linked bug on how I chose the tests that belong to this group.

Bug: 369940767
Change-Id: Ia5f5068023b39e3752738f88745df4cb2b1c1b41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126636
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>