[build] Include percentage + elapsed time in Ninja status
Modify the default `NINJA_STATUS` value to include the
completion percentage and the elapsed time. Note that a
Ninja version that contains the fix at fxr/1133774 is
needed for these values to be updated properly during
the build (without that, the status line only changes
when a command completes).

Fixed: 372310735
Change-Id: Ib20a47e333964e0cee6d20062e23476a3f14989b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134092
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: David Fang <fangism@google.com>