[netstack3] Implement Masquerade NAT
Add support for SNAT, which is performed in the LOCAL_INGRESS and EGRESS
hooks.

SNAT is configured for a connection by rewriting the destination fields
of the reply tuple of a conntrack connection, which serves as a guide to
rewrite future packets seen in both directions on the connection.

Currently the only supported type of SNAT is Masquerade NAT, where an
outgoing packet's source IP address is rewritten to an address that is
assigned to the outgoing interface. The source port of the packet is
also rewritten if either (1) the Masquerade action specifies a rewrite
range and/or (2) the packet's existing source port would cause the
rewritten reply tuple to conflict with another tuple in the conntrack
table. If the netstack fails to find a suitable source port, the packet
is dropped.

Bug: 341771631
Test: netstack3-filter-test
      netstack-filter-fidl-tests
Change-Id: I2bc0af2c2db8b14b5e416bbb0e885b82df2796ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125236
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>