[tee] Introduce TA runtime hooks for resource management
This change introduces hooks to be called as prologues and epilogues to
TA_CreateEntryPoint() and TA_DestroyEntryPoint() for each TA,
respectively. These are intended to be part of a private contract with
the runtime to set up and tear down runtime-specific resources for the
TA (e.g., around protocol connection).

These hooks are intended to panic if something goes awry.

Bug: 371213067
Change-Id: Id5ebdcfdd2e769616063ba8f7a395db633b75595
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135497
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>