[components][sandbox] Replace program_output_router with direct lookups
program_output_router is program_output_dict wrapped by a Router. It is
an argument to build_component_sandbox used to fulfill routing requests
from "self".

This layer of indirection is not necessary, Instead, we can lookup the
Router in the program_output_dict directly, and use that.

Also: Add a dictionary test to exercise more of this logic (hopefully
this can catch a reference cycle, if there is one)

Fixed: 361653724

Change-Id: Ifeb32ab0de16e8ab827c860c010fb506679f2168
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106192
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>