[bazel_sdk] Do not depend on C++ toolchain for preprocess_file()
Before this CL, the preprocess_file() function required the rule
to depend on the current C++ toolchain, which made it impossible
to run in the host build configuration when no such toolchain
is defined.

This CL changes the function to use @fuchsia_clang//:bin/clang-cpp
directly instead,

On the other hand, the dependencies of fuchsia_devicetree_source()
and fuchsia_devicetree_fragment() must be reachable through a
transition to a Fuchsia-compatible build configuration, as these
can be cc_librrary() targets that need a valid C++ toolchain
definition.

Change-Id: Ie88559d5aff80718513f467bf5dc1b6ec4746cd4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135772
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>