[ffx][repo] Use correct repo name with product bundles
This uses the generated name from a server that is serving
a product bundle. Before this change, the target register
command could not find the repo server instance that is
serving a product bundle since the name is prefixed with
the repo_name option, and uses the alias as the suffix of
the name.

Bug: 372026735
Change-Id: I65ef58ba97e70e8887a937187e5e0317b0c373e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133672
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>