[sdk][driver] Add ScopedGlobalLogger to partner
Adds the ScopedGlobalLogger to the sdk. This library is separate from
the regular testing library as it is meant to be used in tests that
don't require setting up the entire driver, which is what the regular
library helps with.

Bug: 369120118

Change-Id: I141a4997890301a376309726bc7134197d6665e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127472
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>