[fuchsia-async] Fix timer memory issue
Prior to this CL, defunct timers wouldn't get removed from the heap
until the original time expired which meant it was possible to end up
with excessive memory use under certain usage patterns.

This change now uses a heap where elements can be removed.  There are
also a number of other optimisations and changes:

1. Timers will now usually avoid heap allocations. A heap allocation
   might be required if the heap needs to grow in capacity, but that
   should not happen very often.  A side-effect of this is that timers
   are no longer Unpin, and timers aren't registered until they're first
   polled.

2. Timer::reset now works no matter what state the timer is in.

3. Timer now implements FusedFuture.

Note that an earlier version of this CL tried to use BTreeSet, but that
caused significant code size increases.

I made some of the callsites Box::pin where it appeared to be difficult
to pin to the stack (which won't be a regression because prior to this
CL all timers caused a heap allocation).  Owners might want to revisit
this to remove the boxing.

Bug: 370598401
Bug: 371152025
Bug: 371434567
Change-Id: I0532bbf2f36ad13bacf2a87ed9141345c31aff0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130312
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Owners-Override: Adam Barth <abarth@google.com>