[debugger] Add DebuggedJob
Add a new DebuggedJob abstraction analogous to DebuggedProcess. This
class will be responsible for monitoring for job exceptions and serving
as an intermediary between the platform specific JobHandle
implementations and DebugAgent.

An additional method is added to JobExceptionObserver which will be used
when we have support for attaching to the non-debugger job exception
channel.

Also adds a new flag to ExceptionNotification to let the front end know
that this is an exception from a Job.

This is not hooked up yet, so no behavior changes are expected.

Change-Id: I0409b1f4f50789c063e9d74f3e4dc59b263dd45b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134738
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>