[starnix][aio] Fix AIOTest.CloneVm
This test was flaky because we were trying to hold onto the memory for
the asynchronous IO operations with a WeakRef to a Task. In this test,
the task that submits the operation can exit before the operation
occurs. Instead, we need to hold a strong reference to the MemoryManager
for the task, which can persist throughout the operation.

Test: AIOTest.CloneVm
Fixed: b/372774866
Change-Id: I1bd4617618c84c184dfe1f0740849527ff86e4da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136332
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>