[starnix][functionfs] Handle R/W on separate thread
Handle the ADB calls on a separate thread so we can use an
async context. The async context is important so we only go
to sleep after we have already queued a read, so we can be
sure we will be woken up when the read returns.

Change-Id: Ibf9114f26d6c6335b1107cc089e29f9ef3100288
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136173
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>