[inspect][ctf] Use default publish settings in Inspect tests
This makes the CTF Inspect publisher test use default PublishOptions.

Sometime ago it seemed better to me to not encode our defaults into CTF
tests. However, with the addition of name-filter-lists in selectors, I
now think it is better to have CTF ensure that if you
  1) Publish using the default Inspect tree name
you will always be able to
  2) Select from Archivist without specifying a name filter list.

Before the use of name-filter-lists in selectors, the tree name was
mostly an internal detail. Now it is part of the selector API, so
ensuring these behaviors stay consistent is important.

Bug: 355732696
Change-Id: I7d00b9d7f9b9e625b3833faac2d99defc75e3461
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129876
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>