[netstack3] Emit the default rule through watchers
Before this change, we never emit the default rule that points to the
main table. This is hard to hold when syncing rules to Core. Instead
of hiding it, we expose this rule at the lowest priority.

Also made a small change to net-cli so that the rules are sorted by
the priority and index.

Bug: 357858471

Change-Id: I843c43d41432808d5e54971567bbf574923aeef1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132379
API-Review: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>