[bazel_sdk][tests] Remove --incompatible_enable_cc_toolchain_resolution=false hack.
This Bazel option was used to allow running the Bazel SDK test suite,
as well as early versions of the platform build, on a system where
there is no GCC or Clang installed (e.g. when running on CI bot
containers).

Unfortunately, support for disabling the feature will be removed
in upcoming versions of Bazel, so remove this by replacing it by
a completely different work-around: define an "empty" C++
toolchain and register it as compatible with the host/exec
build configuration.

Note that BAZEL_DO_NOT_AUTO_DETECT_CPP_TOOLCHAIN=1 still needs
to be used, otherwise hard-coded Bazel repository rules will still
try to probe for Clang and GCC and fail with an error message.

+ Remove --incompatible_enable_cc_toolchain_resolution from
  the build's internal Bazel workspace since this option is now
  the default since Bazel 7.0.

Fixed: 321637402
Change-Id: I9e8d489b6c343621c9f0ec6a0391a1671cf0d887
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135893
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>