[fshost] Move route to correct location
Routes need to be included in the shard that includes the component
using the capability, not the one that declares the component which
provides it. Otherwise we are always forced to include client component
whenever we include theserving component.

This particular route broke microfuchsia as it doesn't include fshost,
but does include the driver framework.

Fixed: 373387565
Change-Id: Id48e7a633ae8ed8a5ebe4c700336e992996013d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137333
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>