[tee][api] Base handle types on u64
Since it will be convenient to model object handle's in FIDL (for the
purposes of a persistent storage service), it is more convenient to base
handle types on u64 over usize (as there is no FIDL integral type with a
Rust binding of usize). This change is cosmetic in practice, as usize
and u64 have the same ABI on all supported architectures - and the
ABI compatibility continues to be verified in the existing tests.

Bug: 371213067
Change-Id: I1454f8ad3484ee1e62cf34f430e8b9a7ab810748
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134731
Reviewed-by: Eric Stone <ecstone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>