[debugger] Move exception record to Exception.
Previously the function to get the debug_ipc::ExceptionRecord was on the
ThreadHandle. This matches the way the Zircon objects work where this is
a property of the "thread" kernel object.

But the exception record is conceptually more associated with the
exception object. This change may possibly reduce race conditions
because the exception record will be queried more closely in time to the
exception delivery. More relevant is that I plan in a followup to add
more POSIX exception information to the exception record which belongs
on the exception object.

There should be no behavior change.

Change-Id: I033dce35c3f5b044c290ce198522eabedd18412c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132532
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>