[starnix][aio] Respect max_operations
Previously, we were counting the maximum number of operations for read
and write operations separately. This CL restructures the internals of
the AIO structures so that we apply the max to the sum of the number of
read and write operations.

Change-Id: Ib5207b91e1814255e6f458ee074134257aa6d81b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137199
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Song <wintermelons@google.com>