[starnix][ltp] Implement CLONE_PARENT
This CL implements the CLONE_PARENT flag for clone, which is tested by
clone08. We now pass an additional subtest of clone08, but we do not
pass the whole test yet because we fail a subtest of another feature.

Test: clone08
Change-Id: I212822137468c775aa24337dbc58090659630c18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137813
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Perry <adamperry@google.com>