[arm][timer] Change the barriers used in sync'ed counter reads.
Change the way that the barriers work when we are attempting to read
the system reference counter and need to make sure that our counter
read takes place after all previous memory loads, but not necessarily
stores.

Prior to this change, if the sync flags had requested that we read
after previous loads, but not stores, we would simply issue an ISB
before the read.  Further review of the most recent ARM ARM implies
that this is not sufficient.  This is shown in example D12-3, which is
a sequence provided to ensure that a counter read takes place _after
receiving a signal via shared memory_.  To make this happen, the
example reads the shared memory variable (call it `X`), and then
issues a branch which depends on the value read, followed by an ISB
before the counter read.  The implication of a very careful read of
the text here is that we are guaranteed to have our counter read take
place after the load of variable `X`.  It does _not_ say that the
counter read will take place after _all_ previous loads.

If a user wants to make certain that the counter load takes place
after _all memory accesses_, the ARM ARM provides example D12-4
instead, which involves using either a DSB, or a DMB coupled with a
creatively constructed artificial dependency.

 Since the (current) spec for
`platform_current_raw_ticks_synchronized` says that the
`kAfterPreviousLoads` flag will ensure that the counter load takes
place after _all_ previous loads, following the example given in D12-3
is not sufficient, and the example given in D12-4 should be used
instead.

There is an opportunity in the future to take advantage of the
(presumably) less expensive sequence given in D12-3, but it would
require a somewhat different API, which provided some way for code
which needed the counter load to take place after the load of a
specific variable `X` to somehow inject a dependency on `X` into the
sequence.  Until then, however, we use the heavier (but actually
correct) hammer.

Change-Id: I219d84c0153ed609f6198b1a56146b5ea03efc01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135072
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: John Grossman <johngro@google.com>