[driver_manager] Parameterize DriverRunnerTest on using dynamic linking.
We want to run each test case on the dynamic linking path, and the
legacy path.

DriverRunnerTest2 is added as a temporary fixture so we can make the
fixture changes required over several CLs. This is as calling GetParam()
will crash if a test has not been migrated to TEST_P, but we don't
want to use the dynamic linking path if part of the test is not yet
supported.

Added the driver-runner-test.cml, so we can add the ability to
create processes that is required by the dynamic linker code.

MULTIPLY: driver-runner-test.cm : 50

Bug: 363012744
Change-Id: I4bc94ecabbc0d78441c16db8fea0aec642e1cc7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136298
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>