[starnix][ltp] Pass utsname03_clone
This test requires the CLONE_NEWUTS flag. We actually had all the logic
for this flag, but we didn't list the flag as supported, which meant we
returned ENOSYS and failed the test.

In addition to adding CLONE_NEWUTS to the list of supported flags, this
CL cleans up the implementation to match the normal naming conventions
for cloning and forking these kinds of objects.

Test: utsname03_clone
Change-Id: I7323a63d69a017ebbc3810ffa995277d9ffc6890
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137816
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>