[terminal] Move packages from legacy bundle to assembly product definition
This moves many of the packages listed directly in the gni file
to the targets in the build.gn file, nearly removing all of the
packages from the cache packages set.  The remaining packages
are in groups that will each be moved in a single cl for clarity.

Run-All-Tests: true

Bug: b/361348304
Change-Id: I2e999b23463a24a3961847efdd1beed0d70eeb04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136090
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>