[debugger] Use DebuggedJob for RootJob
This converts the handling of RootJob to be considered just another one
of the jobs that DebugAgent can attach to, and adds a few methods that
makes it easier to generally attach to jobs using the new DebuggedJob
object.

All notifications that we rely on coming from the root job should be
unchanged.

Change-Id: Ic891a837e9ec51721aa69ee07126d2911f4b9972
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136252
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>