[windowed-stats] Implement event reactor combinators.
This change introduces the `event` module and an API for constructing
reactors that respond to data and system events. Reactors are
constructed with combinators that respond differently to events, such as
sampling user data with a `TimeMatrix`. See unit tests for examples.

To support this change, `TimedSample` has been generalized and renamed
to `Timed` (there are now timed samples and timed events). The
`DataSemantic` trait now has an associated `Metadata` type. Metadata
APIs are incomplete and this change assigns the unit type `()` to
`Metadata` for all `DataSemantic` types.

Bug: 350797382
Test: unit tests
Change-Id: I801bdc5a5f8129781235973d2853e4069368a486
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127883
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>