[starnix] Enable Starnix suspend functionality
This enables the wake_locks feature, which tracks Linux wake locks in
the Starnix kernel. It also delegates suspending the container to the
Starnix runner. The Starnix runner will monitor channels that the
Starnix kernel has registered for incoming messages in order to resume
the container.

Fixed: 370526509
Change-Id: I9ef0c3646f6b56e634c98de371a8c9aa1e3bad9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128713
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>