[amlogic-display] Get MMIO regions using MMIO names.
This replaces GetMmioById() pdev calls with GetMmioByName()
throughout the amlogic-display driver, so that the driver won't rely
on the MMIO region indices defined in the board driver or the
devicetree. This makes it easier for boards to add, remove or change
pdev-provided MMIO resources.

Test: On core.astro, core.sherlock, core.nelson, core.vim3:
      - "display-tool squares" worked.
      - "ffx inspect show" showed correct rdma stats.
      - "flatland_display_compositor_pixeltest" (using capture) passed.
Bug: 372259382
Change-Id: I9f1ae112e4f5eaad836e9aad4fc5fac006ab574b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132519
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>