[driver][sdk] Add ServiceValidator
Adds a new sdk class called ServiceValidator. This class is able to be
provided to the incoming namespace optionally. When provided, the
namespace will perform a validation of the service member connect
operation against the service validator. This can help terminate early
if a driver is connecting to a nonexistent optional node parent, or
connecting to the wrong parent for a specific service.

Bug: 371393843

Change-Id: I7cab6647046047456d1d6724fa688c956920d4d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132775
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>