[starnix] Fix peer-closed logic in proxy
Previously peer closed was not properly handled, and we could end up
with threads just spinning in the wait loop.

The proxy loop code was relying on the message forwarding to return an
error if the read channel was closed, but before reading we checked
whether or not the channel was readable. This meant that the peer closed
was never encountered.

Bug: 371566166
Change-Id: I58ac80cd8579b549be065cb3789221e7bced7298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137133
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>