[gpio][devicetree] Add controller index property for gpio-init nodes
gpio-init composite node specs already have a bind rule for the
controller ID to ensure that they bind to to correct GPIO controller.
This change adds GPIO_CONTROLLER as a bind property as well, using the
index of the GPIO controller as found in the pinctrl-0 property. This
prevents ambiguous bind rules in drivers that need to bind to multiple
gpio-init nodes.

Test: gpioimpl-visitor-test
Change-Id: Ieadd57af3329a252885a72f421105ad2b02acadc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138316
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>