[tee][runtime] Better handle empty VMO parameters
The runtime parameter translation logic from FIDL would unconditionally
map the VMO associated with a memref parameter even when the VMO was
empty, which is illegal.

However, we shouldn't just provide a buffer address of nullptr in such a
case. A zero-length buffer at a valid address is totally fine and that
should be presented to the TA as valid memory. Accordingly, we handle
this case by mapping in a zero-filled page from a fresh VMO - not
trusting that the original VMO had non-zero content - and specify a
buffer of size 0 rooted at that address.

Bug: 371213067
Change-Id: I764c82f4f20cc8f449b5eba8114ded035acf297b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137952
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>